+++ webapp/tests/integration/components/request-error-details/component-test.js
@@ -1,0 +1,24 @@
+import { moduleForComponent, test } from 'ember-qunit';
+import hbs from 'htmlbars-inline-precompile';
+
+moduleForComponent('request-error-details', 'Integration | Component | request error details', {
+  integration: true
+});
+
+test('it renders', function(assert) {
+  // Set any properties with this.set('myProperty', 'value');
+  // Handle any actions with this.on('myAction', function(val) { ... });
+
+  this.render(hbs`{{request-error-details}}`);
+
+  assert.equal(this.$().text().trim(), '');
+
+  // Template block usage:
+  this.render(hbs`
+    {{#request-error-details}}
+      template block text
+    {{/request-error-details}}
+  `);
+
+  assert.equal(this.$().text().trim(), 'template block text');
+});