+++ CHANGES.md
@@ -7,6 +7,7 @@
 * Added the ability to search by product type
 * Added a "status_description" field for tests, allowing setting more informative information on what the test is currently doing (via API)
 * Added timing metrics API, enabling tests and sessions to display the total time distribution spent while running
+* Fixed indication of search term syntax errors
 
 ## Version 2.11.1